Use stronger memory ordering in sync::OnceCell
Although we use once, the `get` call bypasses it, and so it does not
necessary observe all the data written by the initializing function.

In particular, it might be the case that it sees the pointer itself,
but not the actual Box contents. To fix it, switch to the proper
Acquire/Release pair.