implement get_or_try_init for sync::OnceCell
Note that only works with parking_lot, because it requires a mutex
that can be constructed in a const fn

The implementation changes from `Once` to `Mutex`, but this doesn't
affect neither fast-path performance, no datatype size. In theory, we
can do better by adjusting `Mutex` / `Once` implementation to use
their internal flag to store `is_initialized` state, but that is left
to the future work.

The Mutex implementation was inspired by:

  https://github.com/niklasf/double-checked-cell

closes #16