Fix alignment of `imp_std::Waiter`.
Using `#[repr(align(4))]` directly on `Waiter` will downgrade alignment form 8 to 4 bytes on 64-bit targets.

To fix this, an empty struct is defined (`Align4`), which is tagged with `#[repr(align(4))]`, and an instance of this empty struct is added to `Waiter`. `Waiter` will now have an alignment of at least 4 bytes, but it will be greater on targets that require it.