Remove PhantomData<*mut Waiter> so we don't need to manually impl Send
The effect of having `PhantomData<*mut Waiter>` is that the `OnceCell<T>`
stops being `impl<T: Send> Send for OnceCell<T>`. But we want it to be
`impl<T: Send> Send for OnceCell<T>`, because we manually mark it as
such.

Remove the `PhantomData` so we can also remove the manual `impl Send`.